// data/generated/projects.ts
// THIS FILE IS AUTO-GENERATED - DO NOT EDIT MANUALLY
// Generated from: /data/source/master-resume.tex

export interface Project {
  id: number;
  title: string;
  description: string;
  image: string;
  technologies: string[];
  featured: boolean;
  github: string;
}

export const projects: Project[] = [
  {
    "id": 1,
    "title": "Ponotodoro",
    "description": "Developed a full-stack app integrating the Pomodoro technique with note-taking and to-do lists functionality to combine time management and task tracking techniques to improve personal productivity for 10+ users. Engineered frontend with React and Bootstrap, to structure the platform's task management system. Designed backend using Node.js and PostgreSQL for efficient data management and user authentication.",
    "image": "/ponotodoropicture.jpg",
    "technologies": [
      "JavaScript",
      "React",
      "Node.js",
      "Next.js",
      "PostgreSQL",
      "Bootstrap",
      "HTML/CSS"
    ],
    "featured": false,
    "github": "https://github.com/ariqmuldi/ponotodoro"
  },
  {
    "id": 2,
    "title": "Flight Hub",
    "description": "Constructed a full-stack web application that streamlines flight offers by integrating Amadeus and Twilio API. Produced a full-stack blog feature enabling users to create, edit, and manage posts regarding flight-related topics. Created frontend using React and Bootstrap and backend with Flask and SQLite for usability and security.",
    "image": "/flighthubpicture.jpg",
    "technologies": [
      "Python",
      "JavaScript",
      "Flask",
      "React",
      "SQLite",
      "Bootstrap",
      "HTML/CSS"
    ],
    "featured": false,
    "github": "https://github.com/ariqmuldi/flight-hub"
  },
  {
    "id": 3,
    "title": "ChatterBox",
    "description": "Developed a full-stack chat application replicating core Discord functionalities, including real-time messaging, user authentication, and channel management, facilitating communication for 5+ active users across 5+ channels. Engineered Discord-like frontend using React and Tailwind CSS to enhance user experience and responsiveness. Implemented backend with Firebase managing 100+ messages, user data, channel info, and communication.",
    "image": "/chatterboxpicture.jpg",
    "technologies": [
      "JavaScript",
      "React",
      "Firebase",
      "Tailwind CSS",
      "HTML/SCSS"
    ],
    "featured": false,
    "github": "https://github.com/namekeptanonymous/Error404"
  },
  {
    "id": 4,
    "title": "MoodiJawoodi",
    "description": "Produced responsive interface with HTML/CSS and Java with the JDBC API to enhance shopping experience. Designed backend system with Java and MySQL, optimizing data handling and processing of products.",
    "image": "/moodijawoodipicture.jpg",
    "technologies": [
      "Java",
      "HTML/CSS",
      "MySQL",
      "Docker"
    ],
    "featured": false,
    "github": "https://github.com/ariqmuldi/moodi-jawoodi-grocery"
  }
];